backend/requirements.txt
azure-ai-translation-text==1.0.0b1
azure-core==1.29.5
azure-cosmos==4.5.1
azure-functions==1.17.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
flask>=3.0.0
tinydb==4.8.0
typing_extensions==4.8.0
urllib3==2.0.7
uuid==1.30
